tests/wpt/tests/css/css-display/reading-flow/tentative/reading-order-invalid.html

<!DOCTYPE html>
<title>CSS Display: parsing reading-order with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-display-4/#reading-flow">
<link rel="author" title="Di Zhang" href="mailto:dizhangg@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<body>
<script>
test_invalid_value('reading-order', 'auto');
test_invalid_value('reading-order', 'none');
test_invalid_value('reading-order', 'legacy');
test_invalid_value('reading-order', 'normal auto');
test_invalid_value('reading-order', 'flex-visual');
test_invalid_value('reading-order', '1.5');
test_invalid_value('reading-order', '1 2');
test_invalid_value('reading-order', '100%');
test_invalid_value('reading-order', '10px');
</script>
</body>
